add analysis data table
