+++ src/app/waste-water/waste-water-analysis/waste-water-analysis.component.html
@@ -1,2 +1,2 @@
 <app-waste-water-graphs *ngIf="analysisTab == 'graphs'" [containerHeight]="containerHeight"></app-waste-water-graphs>
-<app-waste-water-calculations-table *ngIf="analysisTab == 'tables'"></app-waste-water-calculations-table>
+<app-waste-water-calculations-table *ngIf="analysisTab == 'tables'" [containerHeight]="containerHeight"></app-waste-water-calculations-table>
