+++ src/app/waste-water/waste-water-analysis/waste-water-analysis.component.ts
@@ -24,7 +24,12 @@ export class WasteWaterAnalysisComponent implements OnInit {
     });
     let wasteWater: WasteWater = this.wasteWaterService.wasteWater.getValue();
     let settings: Settings = this.wasteWaterService.settings.getValue();
-    this.wasteWaterAnalysisService.setGraphGata(wasteWater, settings);
+    this.wasteWaterAnalysisService.setResults(wasteWater, settings);
+    this.wasteWaterAnalysisService.setGraphGata();
+    this.wasteWaterAnalysisService.selectedTableData.next({
+      name: 'Baseline',
+      results: this.wasteWaterAnalysisService.baselineResults
+    });
   }
 
   ngOnDestroy() {
