+++ src/app/waste-water/waste-water-banner/analysis-menu/analysis-menu.component.html
@@ -10,20 +10,47 @@
         </div>
     </div>
 
-    <div class="d-flex align-items-center tabs subtabs waste-water ml-auto br-color-white">
+    <div *ngIf="analysisTab == 'graphs'"
+        class="d-flex align-items-center tabs subtabs waste-water ml-auto br-color-white">
         <div class="p-2">
             <div class="nav-item dropdown">
-                <button type="button" class="btn btn-primary w-100" (click)="toggleDropdown()">
+                <button type="button" class="btn btn-primary w-100" (click)="toggleGraphDropdown()">
                     <a class="dropdown-toggle">
                         Select Graphs
                     </a>
                 </button>
-                <div class="dropdown-menu p-0 w-100" [ngClass]="{'show': showDropdown == true}">
+                <div class="dropdown-menu p-0 w-100" [ngClass]="{'show': showGraphDropdown == true}">
                     <div class="list-group">
                         <a *ngFor="let analysisGraphItem of analysisGraphItems"
                             class="list-group-item list-group-item-action" (click)="toggleSelected(analysisGraphItem)"
                             [ngClass]="{'active': analysisGraphItem.selected}">
                             <span [innerHTML]="analysisGraphItem.title"></span>
+                        </a>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <div *ngIf="analysisTab == 'tables'"
+        class="d-flex align-items-center tabs subtabs waste-water ml-auto br-color-white">
+        <div class="p-2">
+            <div class="nav-item dropdown">
+                <button type="button" class="btn btn-primary w-100" (click)="toggleTableDropdown()">
+                    <a class="dropdown-toggle">
+                        Select Scenario: {{selectedTableData.name}}
+                    </a>
+                </button>
+                <div class="dropdown-menu p-0 w-100" [ngClass]="{'show': showTableDropdown == true}">
+                    <div class="list-group">
+                        <a class="list-group-item list-group-item-action"
+                            (click)="selectTableData('Baseline', baselineResults)"
+                            [ngClass]="{'active': selectedTableData.name == 'Baseline'}">Baseline
+                        </a>
+                        <a *ngFor="let modificationsResult of modificationsResultsArr"
+                            class="list-group-item list-group-item-action"
+                            (click)="selectTableData(modificationsResult.name, modificationsResult.results)"
+                            [ngClass]="{'active': selectedTableData.name == modificationsResult.name}">{{modificationsResult.name}}
                         </a>
                     </div>
                 </div>
