+++ src/app/waste-water/waste-water.service.ts
@@ -82,6 +82,7 @@ export class WasteWaterService {
         EnergyCostUnit: aeratorPerformanceCopy.EnergyCostUnit
       }
       let wasteWaterResults: WasteWaterResults = wasteWaterAddon.WasteWaterTreatment(inputData);
+      console.log(wasteWaterResults.SolidsRetentionTime);
       if (settings.unitsOfMeasure != 'Imperial') {
         wasteWaterResults = this.convertWasteWaterService.convertResultsToMetric(wasteWaterResults);
       }
